Bug in select inputs with selected attribute
Hello Dary, 
First of all thanks for your tutorials on YouTube, they are awesome, keep going! :)

I've found out that if you choose an option from any of the select inputs, and the selected option is placed above the default (EUR in select 'to' or USD in select 'from') and submit the form, the result is OK, but selected option is not read from the session value, it is the default (EUR or USD) option selected.

I've submitted quick fix for this issue.

Best Regards.